moved functions relating to config from met_util to config_metplus, ci-run-all-cases
