+++ metplus/wrappers/mtd_wrapper.py
@@ -15,6 +15,7 @@ import os
 from ..util import met_util as util
 from ..util import time_util
 from ..util import do_string_sub
+from ..util import parse_var_list
 from . import CompareGriddedWrapper
 
 class MTDWrapper(CompareGriddedWrapper):
@@ -124,9 +125,9 @@ class MTDWrapper(CompareGriddedWrapper):
             self.read_field_values(c_dict, 'OBS', 'OBS')
 
         c_dict['VAR_LIST_TEMP'] = (
-            util.parse_var_list(self.config,
-                                data_type=c_dict.get('SINGLE_DATA_SRC'),
-                                met_tool=self.app_name)
+            parse_var_list(self.config,
+                           data_type=c_dict.get('SINGLE_DATA_SRC'),
+                           met_tool=self.app_name)
         )
 
         return c_dict
