+++ metplus/wrappers/regrid_data_plane_wrapper.py
@@ -15,6 +15,8 @@ import os
 from ..util import met_util as util
 from ..util import time_util
 from ..util import do_string_sub
+from ..util import parse_var_list
+from ..util import get_process_list
 from . import ReformatGriddedWrapper
 
 # pylint:disable=pointless-string-statement
@@ -106,7 +108,7 @@ class RegridDataPlaneWrapper(ReformatGriddedWrapper):
                 self.log_error("FCST_REGRID_DATA_PLANE_OUTPUT_TEMPLATE must be set if "
                                "FCST_REGRID_DATA_PLANE_RUN is True")
 
-            c_dict['VAR_LIST_FCST'] = util.parse_var_list(
+            c_dict['VAR_LIST_FCST'] = parse_var_list(
                 self.config,
                 data_type='FCST',
                 met_tool=self.app_name
@@ -129,7 +131,7 @@ class RegridDataPlaneWrapper(ReformatGriddedWrapper):
                 self.log_error("OBS_REGRID_DATA_PLANE_OUTPUT_TEMPLATE must be set if "
                                "OBS_REGRID_DATA_PLANE_RUN is True")
 
-            c_dict['VAR_LIST_OBS'] = util.parse_var_list(
+            c_dict['VAR_LIST_OBS'] = parse_var_list(
                 self.config,
                 data_type='OBS',
                 met_tool=self.app_name
@@ -155,7 +157,7 @@ class RegridDataPlaneWrapper(ReformatGriddedWrapper):
         # only check if VERIFICATION_GRID is set if running the tool from the process list
         # RegridDataPlane can be called from other tools like CustomIngest, which sets the
         # verification grid itself
-        if 'RegridDataPlane' in util.get_process_list(self.config):
+        if 'RegridDataPlane' in get_process_list(self.config):
             if not c_dict['VERIFICATION_GRID']:
                 self.log_error("REGRID_DATA_PLANE_VERIF_GRID must be set.")
